add room, work with main page

+++ app/src/main/java/com/ostanets/todoapp/domain/DeleteNoteUseCase.kt
@@ -2,7 +2,7 @@ package com.ostanets.todoapp.domain
 
 class DeleteNoteUseCase(private val noteRepository: NoteRepository) {
 
-    suspend fun deleteNote(noteId: Int) {
+    suspend fun deleteNote(noteId: Long) {
         noteRepository.deleteNote(noteId)
     }
 }
