+++ app/src/main/java/com/ostanets/todoapp/domain/GetNoteUseCase.kt
@@ -1,8 +1,10 @@
 package com.ostanets.todoapp.domain
 
+import com.ostanets.todoapp.models.Note
+
 class GetNoteUseCase(private val noteRepository: NoteRepository) {
 
-    suspend fun getNote(noteId: Int): Note {
+    suspend fun getNote(noteId: Long): Note {
         return noteRepository.getNote(noteId)
     }
 }
